+++ virtual_drive/core.py
@@ -13,6 +13,7 @@ from numpy.typing import NDArray
 from scipy import signal
 from typing_extensions import override
 
+from ingenialink.configuration_file import ConfigurationFile
 from ingenialink.constants import ETH_BUF_SIZE, MONITORING_BUFFER_SIZE
 from ingenialink.dictionary import Interface
 from ingenialink.enums.register import RegAccess, RegDtype
@@ -1423,18 +1424,16 @@ class VirtualDrive(Thread):
         configuration_file = os.path.join(
             pathlib.Path(__file__).parent.resolve(), self.PATH_CONFIGURATION_RELATIVE
         )
-        _, registers = EthernetServo._read_configuration_file(configuration_file)
-        cast_data = {"float": float, "str": str}
-        for element in registers:
-            subnode = int(element.attrib["subnode"])
-            reg_dtype = element.attrib["dtype"]
-            reg_data = element.attrib["storage"]
-            if not self.__register_exists(subnode, element.attrib["id"]):
+        conf_file = ConfigurationFile.load_from_xcf(configuration_file)
+        for conf_register in conf_file.registers:
+            subnode = conf_register.subnode
+            reg_data = conf_register.storage
+            if not self.__register_exists(subnode, conf_register.uid):
                 continue
             self.set_value_by_id(
                 subnode,
-                element.attrib["id"],
-                cast_data.get(reg_dtype, int)(reg_data),
+                conf_register.uid,
+                reg_data,
             )
         value: Union[str, int]
         for subnode in self.__dictionary.subnodes: